+++ api_version.go
@@ -7,6 +7,6 @@
 package stripe
 
 const (
-	apiVersion     string = "2022-11-15"
+	apiVersion     string = "2023-08-16"
 	previewVersion string = "2023-08-11.preview-v2"
 )