+++ quote/client.go
@@ -346,19 +346,19 @@ func (i *InvoiceLineItemIter) InvoiceLineItemList() *stripe.InvoiceLineItemList 
 }
 
 // ListPreviewInvoices is the method for the `GET /v1/quotes/{quote}/preview_invoices` API.
-func ListPreviewInvoices(params *stripe.QuoteListPreviewInvoicesParams) *InvoiceIter {
+func ListPreviewInvoices(params *stripe.QuoteListPreviewInvoicesParams) *PreviewInvoiceIter {
 	return getC().ListPreviewInvoices(params)
 }
 
 // ListPreviewInvoices is the method for the `GET /v1/quotes/{quote}/preview_invoices` API.
-func (c Client) ListPreviewInvoices(listParams *stripe.QuoteListPreviewInvoicesParams) *InvoiceIter {
+func (c Client) ListPreviewInvoices(listParams *stripe.QuoteListPreviewInvoicesParams) *PreviewInvoiceIter {
 	path := stripe.FormatURLPath(
 		"/v1/quotes/%s/preview_invoices",
 		stripe.StringValue(listParams.Quote),
 	)
-	return &InvoiceIter{
+	return &PreviewInvoiceIter{
 		Iter: stripe.GetIter(listParams, func(p *stripe.Params, b *form.Values) ([]interface{}, stripe.ListContainer, error) {
-			list := &stripe.InvoiceList{}
+			list := &stripe.QuotePreviewInvoiceList{}
 			err := c.B.CallRaw(http.MethodGet, path, c.Key, b, p, list)
 
 			ret := make([]interface{}, len(list.Data))
@@ -371,37 +371,37 @@ func (c Client) ListPreviewInvoices(listParams *stripe.QuoteListPreviewInvoicesP
 	}
 }
 
-// InvoiceIter is an iterator for invoices.
-type InvoiceIter struct {
+// PreviewInvoiceIter is an iterator for quote preview invoices.
+type PreviewInvoiceIter struct {
 	*stripe.Iter
 }
 
-// Invoice returns the invoice which the iterator is currently pointing to.
-func (i *InvoiceIter) Invoice() *stripe.Invoice {
-	return i.Current().(*stripe.Invoice)
+// QuotePreviewInvoice returns the quote preview invoice which the iterator is currently pointing to.
+func (i *PreviewInvoiceIter) QuotePreviewInvoice() *stripe.QuotePreviewInvoice {
+	return i.Current().(*stripe.QuotePreviewInvoice)
 }
 
-// InvoiceList returns the current list object which the iterator is
+// QuotePreviewInvoiceList returns the current list object which the iterator is
 // currently using. List objects will change as new API calls are made to
 // continue pagination.
-func (i *InvoiceIter) InvoiceList() *stripe.InvoiceList {
-	return i.List().(*stripe.InvoiceList)
+func (i *PreviewInvoiceIter) QuotePreviewInvoiceList() *stripe.QuotePreviewInvoiceList {
+	return i.List().(*stripe.QuotePreviewInvoiceList)
 }
 
 // ListPreviewSubscriptionSchedules is the method for the `GET /v1/quotes/{quote}/preview_subscription_schedules` API.
-func ListPreviewSubscriptionSchedules(params *stripe.QuoteListPreviewSubscriptionSchedulesParams) *SubscriptionScheduleIter {
+func ListPreviewSubscriptionSchedules(params *stripe.QuoteListPreviewSubscriptionSchedulesParams) *PreviewScheduleIter {
 	return getC().ListPreviewSubscriptionSchedules(params)
 }
 
 // ListPreviewSubscriptionSchedules is the method for the `GET /v1/quotes/{quote}/preview_subscription_schedules` API.
-func (c Client) ListPreviewSubscriptionSchedules(listParams *stripe.QuoteListPreviewSubscriptionSchedulesParams) *SubscriptionScheduleIter {
+func (c Client) ListPreviewSubscriptionSchedules(listParams *stripe.QuoteListPreviewSubscriptionSchedulesParams) *PreviewScheduleIter {
 	path := stripe.FormatURLPath(
 		"/v1/quotes/%s/preview_subscription_schedules",
 		stripe.StringValue(listParams.Quote),
 	)
-	return &SubscriptionScheduleIter{
+	return &PreviewScheduleIter{
 		Iter: stripe.GetIter(listParams, func(p *stripe.Params, b *form.Values) ([]interface{}, stripe.ListContainer, error) {
-			list := &stripe.SubscriptionScheduleList{}
+			list := &stripe.QuotePreviewScheduleList{}
 			err := c.B.CallRaw(http.MethodGet, path, c.Key, b, p, list)
 
 			ret := make([]interface{}, len(list.Data))
@@ -414,21 +414,21 @@ func (c Client) ListPreviewSubscriptionSchedules(listParams *stripe.QuoteListPre
 	}
 }
 
-// SubscriptionScheduleIter is an iterator for subscription schedules.
-type SubscriptionScheduleIter struct {
+// PreviewScheduleIter is an iterator for quote preview schedules.
+type PreviewScheduleIter struct {
 	*stripe.Iter
 }
 
-// SubscriptionSchedule returns the subscription schedule which the iterator is currently pointing to.
-func (i *SubscriptionScheduleIter) SubscriptionSchedule() *stripe.SubscriptionSchedule {
-	return i.Current().(*stripe.SubscriptionSchedule)
+// QuotePreviewSchedule returns the quote preview schedule which the iterator is currently pointing to.
+func (i *PreviewScheduleIter) QuotePreviewSchedule() *stripe.QuotePreviewSchedule {
+	return i.Current().(*stripe.QuotePreviewSchedule)
 }
 
-// SubscriptionScheduleList returns the current list object which the iterator is
+// QuotePreviewScheduleList returns the current list object which the iterator is
 // currently using. List objects will change as new API calls are made to
 // continue pagination.
-func (i *SubscriptionScheduleIter) SubscriptionScheduleList() *stripe.SubscriptionScheduleList {
-	return i.List().(*stripe.SubscriptionScheduleList)
+func (i *PreviewScheduleIter) QuotePreviewScheduleList() *stripe.QuotePreviewScheduleList {
+	return i.List().(*stripe.QuotePreviewScheduleList)
 }
 
 func getC() Client {